refactored PriceFormatter to swift

+++ DayTrades/Pick.h
@@ -16,10 +16,10 @@
 @property (strong, nonatomic) Account *account;
 @property (strong, nonatomic) NSString *dayOfTrade;
 @property (strong, nonatomic) NSString *symbol;
-@property (assign, nonatomic, readonly) float open;
-@property (assign, nonatomic, readonly) float close;
-@property (assign, nonatomic, readonly) float value;
-@property (assign, nonatomic, readonly) float change;
+@property (assign, nonatomic, readonly) double open;
+@property (assign, nonatomic, readonly) double close;
+@property (assign, nonatomic, readonly) double value;
+@property (assign, nonatomic, readonly) double change;
 @property (assign, nonatomic, readonly) BOOL processed;
 
 + (NSString *)parseClassName;
